requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf_spectacular==0.27.1
freezegun==0.3.4
pytest==8.0.2
django-filter==23.5
drf_api_logger==1.1.15
